requirements.txt
PyYAML~=6.0.3
pygame
numpy
pyyaml
